+++ src/main/java/ru/yandex/practicum/filmorate/dal/storage/film/FilmStorage.java
@@ -8,7 +8,7 @@ public interface FilmStorage {
 
     Film createFilm(Film film);
 
-    Film getFilmById(long id);
+    Film getFilmById(Integer id);
 
     Collection<Film> getAllFilms();
